+++ NEWS.md
@@ -27,6 +27,7 @@
 * Adds an array to hold additional call numbers on the holding and item level ([MODINVSTOR-1432](https://folio-org.atlassian.net/browse/MODINVSTOR-1432))
 * Update `get_items_and_holdings_view` to properly handle inactive locations ([MODINVSTOR-1438](https://folio-org.atlassian.net/browse/MODINVSTOR-1438))
 * Addition of the order field to the existing item-related APIs ([MODINVSTOR-1439](https://folio-org.atlassian.net/browse/MODINVSTOR-1439))
+* Calculate the order field on item create/update ([MODINVSTOR-1443](https://folio-org.atlassian.net/browse/MODINVSTOR-1443))
 
 ### Bug fixes
 * Fix ordering of electronic access items for inventory-hierarchy, oai-pmh-view ([MODINVSTOR-1224](https://folio-org.atlassian.net/browse/MODINVSTOR-1224))